requirements.txt
streamlit>=1.33
pandas>=2.0
openpyxl>=3.1
PyYAML>=6.0
altair>=5.0
